Revert "[Tizen](Vector) Support desired size"
This reverts commit ccde201431dc3525786ce0220902c5208777e1a5.

+++ dali-toolkit/internal/visuals/image/image-visual.cpp
@@ -318,7 +318,7 @@ void ImageVisual::DoSetProperty(Property::Index index, const Property::Value& va
 
     case Toolkit::ImageVisual::Property::DESIRED_WIDTH:
     {
-      int32_t desiredWidth = 0;
+      float desiredWidth = 0.0f;
       if(value.Get(desiredWidth))
       {
         mDesiredSize.SetWidth(desiredWidth);
@@ -332,7 +332,7 @@ void ImageVisual::DoSetProperty(Property::Index index, const Property::Value& va
 
     case Toolkit::ImageVisual::Property::DESIRED_HEIGHT:
     {
-      int32_t desiredHeight = 0;
+      float desiredHeight = 0.0f;
       if(value.Get(desiredHeight))
       {
         mDesiredSize.SetHeight(desiredHeight);
